requirements.txt
numpy
pandas
scipy
scikit-allel
mygene
gseapy